dev-requirements.txt
flake8
pycrypto==2.6.1
pytest==2.6.4
pytest-cov==1.8.1
requests==2.5.1
responses==0.3.0
wheel
twine
